requirements.txt
kafka-python
pyspark
streamlit
pandas
pyarrow
